refactor: extract KgsWidgetUpdater
